+++ lib/convert.h
@@ -42,10 +42,10 @@ char convert_digit_to_character(const char digit);
  * @brief Convert a string to a number.
  *
  * @param string
- * @return long long
+ * @return int64_t
  * @since v1.0.0
  */
-long long convert_string_to_number(const string_t string);
+int64_t convert_string_to_number(const string_t string);
 
 /**
  * @brief Convert a number to a string.
@@ -54,7 +54,7 @@ long long convert_string_to_number(const string_t string);
  * @return string_t
  * @since v1.0.0
  */
-string_t convert_number_to_string(const long long integer);
+string_t convert_number_to_string(const int64_t integer);
 
 /**
  * @brief Convert a number (base 10) to a string with a specific base.
@@ -64,17 +64,17 @@ string_t convert_number_to_string(const long long integer);
  * @return string_t
  * @since v1.0.0
  */
-string_t convert_number_from_base_10_to_base(unsigned long long number, unsigned long base);
+string_t convert_number_from_base_10_to_base(uint64_t number, uint64_t base);
 
 /**
  * @brief Convert a number with a specific base to a number base 10.
  *
  * @param number
  * @param base
- * @return int
+ * @return uint64_t
  * @since v1.0.0
  */
-unsigned long convert_number_from_base_to_base_10(string_t number, unsigned long base);
+uint64_t convert_number_from_base_to_base_10(string_t number, uint64_t base);
 
 /**
  * @brief Convert a number with a specific base to a number of specific base.
@@ -85,6 +85,6 @@ unsigned long convert_number_from_base_to_base_10(string_t number, unsigned long
  * @return string_t
  * @since v1.0.0
  */
-string_t convert_number_from_base_to_another(string_t number, unsigned long base_from, unsigned long base_target);
+string_t convert_number_from_base_to_another(string_t number, uint64_t base_from, uint64_t base_target);
 
 #endif
